refactor: 폴더명 변경 및 build 경고 제거
