+++ app/(route)/activity/details/_components/Calendar/TabletCalendar.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import Image from 'next/image';
 
-import Button from '@/_components/button';
+import Button from '@/_components/Button';
 
 import TimeSelection from './TimeSelection';
 import type { ActivityData } from './types';
