+++ app/(route)/activity/details/_components/Calendar/TimeSelection.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Button from '@/_components/button';
+import Button from '@/_components/Button';
 
 import type { ActivityData } from './types';
