+++ app/(route)/main/_components/AllActivityLists/ActivityGrid/ActivityGrid.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import Rating from '@/_components/rating';
+import Rating from '@/_components/Rating';
 
 import Spinner from 'public/assets/icons/spinner.svg';
