+++ app/(route)/main/_components/HotActivityLists/HotListsCarousel/index.tsx
@@ -9,7 +9,7 @@ import { useRouter } from 'next/navigation';
 
 import useGetActivities from '@/_hooks/activities/useGetActivities';
 
-import Rating from '@/_components/rating';
+import Rating from '@/_components/Rating';
 
 import usePrevNextButtons, { NextButton, PrevButton } from './HotListsCarouselBtn';
