+++ app/(route)/my/account/_components/AccountConfirm/index.tsx
@@ -15,7 +15,7 @@ import type { ErrorResponse } from '@/_types/userAccount';
 
 import { setCookie } from '@/_utils/cookie';
 
-import Button from '@/_components/button';
+import Button from '@/_components/Button';
 import Input from '@/_components/Input';
 
 import DefaultProfile from 'public/assets/icons/default-profile.svg';
