+++ app/(route)/my/reservations/_components/CancelButton/index.tsx
@@ -8,7 +8,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import useModalState from '@/_hooks/useModalState';
 
-import Button from '@/_components/button';
+import Button from '@/_components/Button';
 
 import CancelModal from '../CancelModal';
