+++ app/(route)/my/reservations/_components/FilterDropdown/index.tsx
@@ -8,8 +8,8 @@ import Image from 'next/image';
 import { useSearchParams } from 'next/navigation';
 
 import type { ReservationParams } from '@/_types/myReservations';
-import type { ReservationFilter } from '@/_constants/reservation-filter';
-import { RESERVATION_FILTER } from '@/_constants/reservation-filter';
+import type { ReservationFilter } from '@/_constants/reservationFilter';
+import { RESERVATION_FILTER } from '@/_constants/reservationFilter';
 
 import useSelectBox from '@/_hooks/useSelectBox';
