+++ app/_components/CalendarNotice/_components/StatusChips/index.tsx
@@ -14,7 +14,7 @@ export interface StatusChipProps {
  */
 export default function StatusChip({ count, bgColor, textColor, label }: StatusChipProps) {
   return (
-    <span className={`mx-0.5 mb-0.5 flex w-full gap-0.5 rounded px-1 text-m font-medium max-mobile:text-s ${bgColor} ${textColor}`}>
+    <span className={`text-m max-mobile:text-s mx-0.5 mb-0.5 flex w-full gap-0.5 rounded px-1 font-medium ${bgColor} ${textColor}`}>
       <p>{label}</p> <p>{count}</p>
     </span>
   );
