+++ app/_components/SideUserProfileCard/_components/ChangeProfileImageModal/index.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 
 import { DEFAULT_IMAGE } from '@/_constants/sideUserProfileCard';
 
-import Modal from '@/_components/modal';
+import Modal from '@/_components/Modal';
 
 interface ChangeProfileImageModalProps {
   currentProfileImage: string;
