--- app/_components/example-jest-test/index.tsx
@@ -1,3 +1,0 @@
-const ExampleJestTest = (a: number, b: number) => a + b;
-
-export default ExampleJestTest;
