--- app/_components/example-zustand/index.tsx
@@ -1,21 +1,0 @@
-'use client';
-
-import { useExampleStore } from '@/_stores/useExampleStore';
-
-export default function ExampleZustand() {
-  const bears = useExampleStore.use.bear();
-  const setBear = useExampleStore.use.setBear();
-  return (
-    <>
-      {bears} around here...
-      <button
-        onClick={() => {
-          setBear((prevBear) => prevBear + 1);
-        }}
-        type="button"
-      >
-        one up
-      </button>
-    </>
-  );
-}
