+++ app/layout.tsx
@@ -5,7 +5,7 @@ import type { Metadata } from 'next';
 import { getLoginStatus } from '@/_utils/isLogin';
 
 import Footer from './_components/Footer';
-import Header from './_components/header';
+import Header from './_components/Header';
 import Providers from './providers';
 
 export const metadata: Metadata = {
