+++ app/not-found.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import Error from 'public/assets/images/404-error.gif';
 
-import Button from './_components/button';
+import Button from './_components/Button';
 
 export default function NotFound() {
   return (
